DjangoBackend/backend/urls.py
"""
URL configuration for project_backend project.

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/5.1/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
# project_backend/urls.py
from django.contrib import admin
from django.urls import path, include
from rest_framework_simplejwt.views import TokenObtainPairView,TokenRefreshView

urlpatterns = [
    path('admin/', admin.site.urls), # Django Admin Panel
    path('django_plotly_dash/', include('django_plotly_dash.urls')), # dash
    path('dash_app/', include('backend.dash_app.urls')),
    path("", include("users.urls")), # Web app home & log-in page
    
    # API Routes for iOS App
    path("api/users/", include("users.api_urls")),  # Separate API for users
    path("api/workouts/", include("workouts.urls")),  # API for workouts
    
    # JWT authentication endpoints
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]
