DjangoBackend/requirements.txt
asgiref==3.8.1
dash==3.0.2
dash-bootstrap-components==2.0.0
Django==5.1.7
django-cors-headers==4.7.0
django-plotly-dash==2.4.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
